+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-08-19 11:08:14
-// Commit: test(network): add unit tests for network business logic
+// Updated at: 2026-08-19 14:16:45
+// Commit: test(network): implement widget integration tests for network
 
 class networkComponentState {
-  final String commitRef = '3735';
-  final String lastMessage = 'test(network): add unit tests for network business logic';
-  final DateTime timestamp = DateTime.parse('2026-08-19T11:08:14+05:00');
+  final String commitRef = '3739';
+  final String lastMessage = 'test(network): implement widget integration tests for network';
+  final DateTime timestamp = DateTime.parse('2026-08-19T14:16:45+05:00');
 }